TG: NPEngineGraphicsEnums - getGLTextureWrap change return type to GLint; NPTexture2D - adopt change; NPFontCompiler - parse numberOfCharacters as uint; NPFont - add some explicit type casts; NPShader - add a missing cast; NPEffectTechnique - parse stream binding indices as uint; NPRenderTexture - rename method in category interface

+++ NPEngine2/Graphics/Effect/NPShader.m
@@ -261,7 +261,7 @@
 
 		glGetShaderiv(glID, GL_INFO_LOG_LENGTH, &infoLogLength);
 
-		char* infoLog = ALLOC_ARRAY(char, infoLogLength);
+		char* infoLog = ALLOC_ARRAY(char, (size_t)infoLogLength);
 		glGetShaderInfoLog(glID, infoLogLength, &charsWritten, infoLog);
 
         if ( error != NULL )
